+++ backend/routes/users.js
@@ -83,4 +83,39 @@ module.exports = function users(app, logger) {
             response.status(500).json({ message: err.message });
         }
     })
+    app.put('/users/updatePassword', async (request, response) => {
+        try {
+            const username = request.body.username
+            const newPass = request.body.password
+            const { createHash } = require('crypto');
+        function hash(string) {
+            return createHash('sha256').update(string).digest('hex');
+        }
+        const hashed = hash(newPass + 'aB6nkeF0He3imq4AOhbO5kEljbveRpLn');
+        const query = knex('users')
+            .where({ username: username })
+            .update({ password: hashed })
+        const results = await query;
+        console.log('Results of my PUT statement: ', results);
+        response.status(200).json({ username: username});
+        } catch (err) {
+            console.error('There was an error in PUT /users/updatePassword', err);
+            response.status(500).json({ message: err.message });
+        }
+    })
+    app.delete('/users', async (request, response) => {
+        try {
+            console.log('Initiating DELETE /users request');
+            console.log('Request has a body / payload containing:', request.body);
+            console.log('Request has params containing:', request.query);
+            const username = request.body.username
+            const query = knex('users').delete().where({username});
+            const results = await query;
+            console.log('Results of my DELETE statement:', results);
+            response.status(200).json(results);
+        } catch (err) {
+            console.error('There was an error in DELETE /users', err);
+            response.status(500).json({ message: err.message });
+        }
+    });
 }
